fix(zscript): split z.py into per-lifecycle mixins
Closes #299.

The 1693-line .nanvix/z.py is split into per-stage mixins under
.nanvix/src/, with z.py retained as the entry point composing them:

  src/lib.py        LibMixin    shared helpers + cross-mixin forward
                                declarations
  src/setup.py      SetupMixin
  src/build.py      BuildMixin  also owns site-packages + ramfs/initrd
                                staging
  src/test.py       TestMixin
  src/release.py    ReleaseMixin
  src/benchmark.py  BenchmarkMixin
  src/clean.py      CleanMixin

NanvixPythonBuild composes them via the MRO; no logic moved between
stages and the body of each lifecycle hook is byte-faithful to the
pre-split z.py modulo the underscore-prefix drop on a handful of
shared names (IS_WINDOWS, DEFAULT_TIMEOUT, nanvixd_binary(),
mkramfs_binary()) that are now imported from src/lib.py.

LibMixin declares NotImplementedError forward-decl stubs for the
cross-mixin helpers (_ensure_initrd, _ensure_ramfs, _install_boot_script,
_install_site_packages, _cleanup_initrd, _cleanup_ramfs) so basedpyright
resolves the cross-mixin calls without Protocol plumbing.

z.py ends with 'del SetupMixin, BuildMixin, ...' as a temporary
workaround: nanvix_zutil's discover_script_class() returns the *first*
ZScript subclass in vars(module) and would otherwise pick a partial
mixin instead of NanvixPythonBuild. The upstream fix is tracked at
nanvix/zutils#269.
