+++ .nanvix/src/__init__.py
@@ -1,0 +1,1 @@
+"""Lifecycle mixins for the nanvix-python ZScript."""
