+++ .nanvix/src/clean.py
@@ -1,0 +1,50 @@
+# Copyright(c) The Maintainers of Nanvix.
+# Licensed under the MIT License.
+
+"""Clean lifecycle for the nanvix-python ZScript."""
+
+from __future__ import annotations
+
+import shutil
+from pathlib import Path
+
+from nanvix_zutil import CFG_SYSROOT, log, paths
+from nanvix_zutil.paths import nanvix_root, repo_root
+
+from .lib import LibMixin
+
+
+class CleanMixin(LibMixin):
+    """``./z clean`` — remove build artifacts."""
+
+    def clean(self) -> None:
+        """Remove build artifacts."""
+        # Clean release assets
+        dist_dir = paths.dist_dir()
+        if dist_dir.is_dir():
+            shutil.rmtree(dist_dir)
+        release_dir = repo_root() / "release-assets"
+        if release_dir.is_dir():
+            shutil.rmtree(release_dir)
+
+        # Clean ramfs artifacts
+        self._cleanup_ramfs()
+        ramfs_img = nanvix_root() / "nanvix_rootfs.img"
+        ramfs_img.unlink(missing_ok=True)
+        ramfs_sentinel = nanvix_root() / ".ramfs-built"
+        ramfs_sentinel.unlink(missing_ok=True)
+
+        # Clean initrd
+        self._cleanup_initrd()
+
+        # Clean snapshot artifacts produced by the snapshot smoke test
+        sysroot_str = self.config.get(CFG_SYSROOT, "")
+        if sysroot_str:
+            snapshots_dir = Path(sysroot_str) / "snapshots"
+            if snapshots_dir.is_dir():
+                shutil.rmtree(snapshots_dir)
+
+        # Clean python3.12 copy used by make_initrd
+        self._cleanup_python_in_repo_root()
+
+        log.success("clean complete")
